Fixed Book update (#200)
* Fixed Book update
Fixed Book validation

* Methods renamed
Unit tests changed

+++ src/Application/Services/Implementation/BookService.cs
@@ -26,11 +26,10 @@ namespace Application.Services.Implementation
         private readonly IUserResolverService _userResolverService;
         private readonly IPaginationService _paginationService;
         private readonly IImageService _imageService;
-        private readonly BookCrossingContext _context;
         private readonly IMapper _mapper;
 
         public BookService(IRepository<Book> bookRepository, IMapper mapper, IRepository<BookAuthor> bookAuthorRepository, IRepository<BookGenre> bookGenreRepository,
-            IRepository<UserLocation> userLocationRepository, IPaginationService paginationService, IRepository<Request> requestRepository, BookCrossingContext context, 
+            IRepository<UserLocation> userLocationRepository, IPaginationService paginationService, IRepository<Request> requestRepository,
             IUserResolverService userResolverService, IImageService imageService)
         {
             _bookRepository = bookRepository;
@@ -39,13 +38,12 @@ namespace Application.Services.Implementation
             _userLocationRepository = userLocationRepository;
             _requestRepository = requestRepository;
             _paginationService = paginationService;
-            _context = context;
             _mapper = mapper;
             _imageService = imageService;
             _userResolverService = userResolverService;
         }
 
-        public async Task<BookGetDto> GetById(int bookId)
+        public async Task<BookGetDto> GetByIdAsync(int bookId)
         {
             return _mapper.Map<BookGetDto>(await _bookRepository.GetAll()
                                                                .Include(p => p.BookAuthor)
@@ -58,7 +56,7 @@ namespace Application.Services.Implementation
                                                                .FirstOrDefaultAsync(p => p.Id == bookId));
         }
 
-        public async Task<BookGetDto> Add(BookPostDto bookDto)
+        public async Task<BookGetDto> AddAsync(BookPostDto bookDto)
         {
             var book = _mapper.Map<Book>(bookDto);
             if (bookDto.Image != null)
@@ -70,40 +68,55 @@ namespace Application.Services.Implementation
             return _mapper.Map<BookGetDto>(book);
         }
 
-        public async Task<bool> Remove(int bookId)
+        public async Task<bool> RemoveAsync(int bookId)
         {
-            var book = await _bookRepository.GetAll()
-                            .FirstOrDefaultAsync(p => p.Id == bookId);
+            var book = await _bookRepository.FindByIdAsync(bookId);
             if (book == null)
+            {
                 return false;
-            _imageService.DeleteImage(book.ImagePath);
+            }
+            if (book.ImagePath != null)
+            {
+                _imageService.DeleteImage(book.ImagePath);
+            }
             _bookRepository.Remove(book);
             var affectedRows = await _bookRepository.SaveChangesAsync();
             return affectedRows > 0;
         }
 
-        public async Task<bool> Update(BookPutDto bookDto)
+        public async Task<bool> UpdateAsync(BookPutDto bookDto)
         {
-            using (var transaction = _context.Database.BeginTransaction())
+            var book = _mapper.Map<Book>(bookDto);
+            var oldBook = await _bookRepository.GetAll().AsNoTracking().FirstOrDefaultAsync(a => a.Id == book.Id);
+            if (oldBook == null)
             {
-                var book = _mapper.Map<Book>(bookDto);
-                var doesBookExist = await _bookRepository.GetAll().AnyAsync(a => a.Id == book.Id);
-                if (!doesBookExist)
-                {
-                    return false;
-                }
-                _bookAuthorRepository.RemoveRange(await _bookAuthorRepository.GetAll().Where(a => a.BookId == book.Id).ToListAsync());
-                _bookGenreRepository.RemoveRange(await _bookGenreRepository.GetAll().Where(a => a.BookId == book.Id).ToListAsync());
-                await _bookRepository.SaveChangesAsync();
-                _bookAuthorRepository.AddRange(book.BookAuthor);
-                _bookGenreRepository.AddRange(book.BookGenre);
-                _bookRepository.Update(book);
-                var affectedRows = await _bookRepository.SaveChangesAsync();
-                transaction.Commit();
-                return affectedRows > 0;
+                return false;
             }
+            if (bookDto.FieldMasks.Contains("Image"))
+            {
+                string imagePath;
+                bookDto.FieldMasks.Remove("Image");
+                bookDto.FieldMasks.Add("ImagePath");
+                if (oldBook.ImagePath != null)
+                {
+                    _imageService.DeleteImage(oldBook.ImagePath);
+                }
+                if (bookDto.Image != null)
+                {
+                    imagePath = await _imageService.UploadImage(bookDto.Image);
+                }
+                else
+                {
+                    imagePath = null;
+                }
+                book.ImagePath = imagePath;
+            }
+            await _bookRepository.Update(book, bookDto.FieldMasks);
+            var affectedRows = await _bookRepository.SaveChangesAsync();
+            return affectedRows > 0;
         }
-        public async Task<PaginationDto<BookGetDto>> GetAll(BookQueryParams parameters)
+
+        public async Task<PaginationDto<BookGetDto>> GetAllAsync(BookQueryParams parameters)
         {
             var query = GetFilteredQuery(_bookRepository.GetAll(), parameters);
             return await _paginationService.GetPageAsync<BookGetDto, Book>(query, parameters);
